+++ mllib/src/main/scala/org/apache/spark/mllib/clustering/topicmodeling/topicmodels/SparseVectorFasterSum.scala
@@ -23,7 +23,7 @@ import breeze.linalg.SparseVector
  * Created by valerij on 6/25/14.
  */
 trait SparseVectorFasterSum {
-  protected def sum[T](vector: SparseVector[Short]) = {
+  protected def sum[T](vector: SparseVector[Int]) = {
     var sum = 0
     for (element <- vector) sum += element
     sum
